feat: fix WASM build for crates dependent on reth-payload-primitives
