+++ crates/chain-state/Cargo.toml
@@ -27,7 +27,7 @@ alloy-primitives.workspace = true
 alloy-consensus.workspace = true
 
 # async
-tokio = { workspace = true, features = ["sync", "macros", "rt-multi-thread"] }
+tokio = { workspace = true, default-features = false, features = ["sync", "macros"] }
 tokio-stream = { workspace = true, features = ["sync"] }
 
 # tracing
