+++ crates/payload/primitives/Cargo.toml
@@ -14,8 +14,11 @@ workspace = true
 [dependencies]
 # reth
 reth-chainspec.workspace = true
+reth-errors.workspace = true
 reth-primitives.workspace = true
 reth-chain-state.workspace = true
+
+revm-primitives.workspace = true
 
 # alloy
 alloy-eips.workspace = true
@@ -26,6 +29,7 @@ op-alloy-rpc-types-engine = { workspace = true, optional = true }
 # misc
 serde.workspace = true
 thiserror.workspace = true
+tokio = { workspace = true, default-features = false, features = ["sync"] }
 
 [features]
 op = ["dep:op-alloy-rpc-types-engine"]
